JS to AMD format, Privacy API implemented, CI/docs updated
