+++ .travis.yml
@@ -1,47 +1,70 @@
+sudo: required
+
 language: php
 
-sudo: false
+dist: xenial
+
+services:
+  - mysql
+
+php:
+  - 7.2
+  - 7.1
+  - 7.0
+
+addons:
+  postgresql: "9.6"
+
+env:
+  global:
+    - MOODLE_BRANCH=MOODLE_37_STABLE
+    - MOODLE_BRANCH=MOODLE_36_STABLE
+    - MOODLE_BRANCH=MOODLE_35_STABLE
+  matrix:
+    - DB=pgsql
+    - DB=mysqli
 
 cache:
   directories:
     - $HOME/.composer/cache
-
-php:
- - 5.6
- - 7.0
-
-matrix:
- allow_failures:
-  - php: 7.0
-
-env:
- matrix:
-  - DB=mysqli MOODLE_BRANCH=MOODLE_30_STABLE
-  - DB=mysqli MOODLE_BRANCH=MOODLE_31_STABLE
-  - DB=mysqli MOODLE_BRANCH=MOODLE_32_STABLE
-  - DB=mysqli MOODLE_BRANCH=MOODLE_33_STABLE
-  - DB=pgsql MOODLE_BRANCH=MOODLE_30_STABLE
-  - DB=pgsql MOODLE_BRANCH=MOODLE_31_STABLE
-  - DB=pgsql MOODLE_BRANCH=MOODLE_32_STABLE
-  - DB=pgsql MOODLE_BRANCH=MOODLE_33_STABLE
-
+    - $HOME/.npm
 
 before_install:
+  - phpenv config-rm xdebug.ini
+  - nvm install 8.9
+  - nvm use 8.9
   - cd ../..
-  - composer selfupdate
-  - composer create-project -n --no-dev moodlerooms/moodle-plugin-ci ci ^1
+  - composer create-project -n --no-dev --prefer-dist blackboard-open-source/moodle-plugin-ci ci ^2
   - export PATH="$(cd ci/bin; pwd):$(cd ci/vendor/bin; pwd):$PATH"
 
 install:
   - moodle-plugin-ci install
+  - >
+    if [ "$DB" = 'mysqli' ];
+    then
+        sudo mkdir /mnt/ramdisk
+        sudo mount -t tmpfs -o size=1024m tmpfs /mnt/ramdisk
+        sudo service mysql stop
+        sudo mv /var/lib/mysql /mnt/ramdisk
+        sudo ln -s /mnt/ramdisk/mysql /var/lib/mysql
+        sudo service mysql restart
+    fi
+  - >
+    if [ "$DB" = 'pgsql' ];
+    then
+        sudo mkdir /mnt/ramdisk
+        sudo mount -t tmpfs -o size=1024m tmpfs /mnt/ramdisk
+        sudo service postgresql stop
+        sudo mv /var/lib/postgresql /mnt/ramdisk
+        sudo ln -s /mnt/ramdisk/postgresql /var/lib/postgresql
+        sudo service postgresql start 9.6
+    fi
 
 script:
   - moodle-plugin-ci phplint
   - moodle-plugin-ci phpcpd
   - moodle-plugin-ci phpmd
   - moodle-plugin-ci codechecker
-  - moodle-plugin-ci csslint
-  - moodle-plugin-ci shifter
-  - moodle-plugin-ci jshint
-  - moodle-plugin-ci phpunit
-  - moodle-plugin-ci behat
+  - moodle-plugin-ci phpdoc
+  - moodle-plugin-ci validate
+  - moodle-plugin-ci savepoints
