+++ styles.css
@@ -19,8 +19,7 @@
 .block_enrolmenttimer .content .timerNumChar {
     -moz-border-radius: 3px;
     -webkit-border-radius: 3px;
-    background-color: #000;
-    background: url([[pix:block_enrolmenttimer|background]]) repeat-x;
+    background: #000 url([[pix:block_enrolmenttimer|background]]) repeat-x;
     border-radius: 3px;
     color: #fff;
     font-size: 1.3em;
